Добавлен миксин CRUDPasswordMixin для управления паролями пользователей, включая методы для сброса пароля, создания токенов восстановления и проверки паролей. Обновлены классы CRUDAdminUser и CRUDModeratorUser для использования нового миксина.
